LayoutTests/platform/gtk/accessibility/loading-iframe-sends-notification-expected.txt

Before


After

End of test

This tests that when an iframe finishes loading, it sends a layout complete notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS subwebarea.childrenCount > 0 is false

TEST COMPLETE
